requirements.txt
python-telegram-bot==21.11.1
python-dotenv==1.0.1
google-api-python-client==2.169.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
